rmi4update: Optimize writing firmware blocks
Most HID devices allow appending the reflash command to the end of the firmware
block. This avoids sending a second report with the just the command. Also, after
the block is written HID devices send an attention report. Only read the F34 control
registers if waiting for that attention report times out.