rmihidtool: Search for a device if non is provided
For simple cases where there is only one Synaptics device this allows
the tool to figure out which device to open without forcing the user
to figure out which device to open ahead of time. If there are more
then one device the user can explicitly specify the device with the -d
option. Which has been repurposed. Printing device into now uses -n.