Fix Application.mk.
Current versions of the NDK only have libc++, and no longer support the
armeabi ABI (as opposed to armeabi-v7a, which is still supported). It's
possible that the APP_ABI line should just be removed, leaving ndk-build
to default to building for all supported ABIs, but for now I've kept it.